This grammar was built off of Peter's grammar and handles everythin gin the tutorial and is not ambiguous (mainly because there are mad lookaheads but who cares).  Here is the commit message
"Pass.g, as far as I can tell, works.  What I fixed: expr now does a lookahead for either assignments or function definitions, if neither then it tries a short statement which are return or break statements.  If it isn't a short_stmt we start parsing it as if it was an expression that can be included in logical operations, like atoms, anonymous lambda functions (ie anon function declarations with arguments that evaluates to something), etc.  If you want to see how the grammar rejects a simple function declaration when in the context of logical operators like + or ||, look at factor: '(' factor_p.  Because we have a production boolean that will always return something we can evaluate (either it's a string or it's a number or an ID with an accessor), the control statements use bool instead of eval for their conditions.  Furthermore, there are no ambiguous paths because I have put in enough lookaheads.  I ran as many pieces of code in the LRM as possible and all passed!  There are some undefined indenting issues we need to work out, but for the most part... this works"